Move AthleteCreationForm to dedicated file
